requirements.txt
flask>=3.0.0
groq>=0.8.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
dnspython>=2.4.0
